Refactor SelectorDialog
- move SelectorDialog to ui/extensions package
- replace on select listener interface by function type
- apply changes to SendHodlerFragment
